+++ app/src/test/snapshots/images/com.adammcneilly.pokedex.pokemonlist_PokemonListCardPaparazziTest_renderTypes[useDarkTheme=false,DARK].png
