+++ app/src/test/snapshots/images/com.adammcneilly.pokedex.pokemonlist_PokemonListCardPaparazziTest_renderTypes[useDarkTheme=true,ROCK].png
